+++ Domibus-MSH/src/main/java/eu/domibus/core/csv/CsvServiceImpl.java
@@ -72,14 +72,14 @@ public class CsvServiceImpl implements CsvService {
     }
 
     @Override
-    public void validateMaxRows(Integer count) throws RequestValidationException {
+    public void validateMaxRows(long count) throws RequestValidationException {
         validateMaxRows(count, null);
     }
 
     @Override
-    public void validateMaxRows(Integer count, Supplier<Long> countMethod) throws RequestValidationException {
+    public void validateMaxRows(long count, Supplier<Long> countMethod) throws RequestValidationException {
         if (count > getMaxNumberRowsToExport()) {
-            Long all = countMethod == null ? Long.valueOf(count) : countMethod.get();
+            Long all = countMethod == null ? count : countMethod.get();
             String message = String.format("The number of elements to export [%s] exceeds the maximum allowed [%s]."
                     , all, getMaxNumberRowsToExport());
             throw new RequestValidationException(message);
